% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaelbow.R
\name{elbowplot}
\alias{elbowplot}
\title{Elbow (Scree) Plot}
\usage{
elbowplot(object, n = "all")
}
\arguments{
\item{object}{a object of class \code{pca} generated using the \code{pca()} function.}

\item{n}{a single numeric value indicating the number of principal components to include in the plot.}
}
\value{
Returns a ggplot. This plot can be modified by adding additional ggplot functions.
}
\description{
Plots the percent variance explained by each principal component from a \code{pca} object.
As a rule of thumb, points above the 'elbow' of the plot should be retained in downstream analyses.
}
\examples{
require(ggplot2)
data(diamonds)
p1 = pca(~carat + depth + table + price+ x + y + z, data=diamonds)
elbowplot(p1)
# modify with additional ggplot functions:
elbowplot(p1) + geom_hline(aes(yintercept=.05), col='red')

elbowplot(p1, n=4) # only plot the first 4 PCs
}
