% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_blog.R
\name{lm_blog}
\alias{lm_blog}
\title{Fit a Linear Regression model}
\usage{
lm_blog(eqn, dataset)
}
\arguments{
\item{eqn}{Equation of the model to be fitted. The argument is of the form :Y ~ X1 + X2... + Xn.}

\item{dataset}{Dataframe containing the covariates and response variable as column names.}
}
\value{
lm_blog returns a list "lr". The list "lr" contains the following components:\cr
\tabular{rrrrr}{
  \strong{Element} \tab \strong{Desciption}\cr
  \emph{dataset} \tab Equation used to fit the model.\cr
  \emph{eqn} \tab Equation used to fit the model.\cr
  \emph{beta_hat} \tab Coefficient estimates.\cr
  \emph{y_hat} \tab Predicted Y values.\cr
  \emph{sigma_2_hat} \tab Estimated model variance.\cr
  \emph{var_hat_b} \tab Variance covariance matrix.\cr
  \emph{e_hat} \tab Equation used to fit the model.\cr
  \emph{dfy} \tab No: of observations - No: of parameters.\cr
  \emph{dfx} \tab No: of parameters - 1.\cr
  \emph{mse} \tab Mean Squared Error.\cr
  \emph{ssy} \tab Total sum of squares.\cr
  \emph{y} \tab Response variable.\cr
  \emph{x} \tab Covariates.
}
}
\description{
lm_blog is used to fit Simple Linear Regression and Multiple Linear Regression models with continous or categorical covariates.
}
\details{
lm_blog takes an equation and a dataframe as arguments to fit a linear regression model represented by the equation.
The dataframe should contain the response variable and the covariates present in the equation as column names.\cr The covariates are treated as continous or categorical. lm_blog automatically converts any non numeric data time into dummy variables using one-got encoding.\cr lm_blog then fits the model using the method of least squares
}
\examples{
library(datasets)\crdata(iris)\crlm_blog(Sepal.Length ~ Sepal.Width + Petal.Length + Species,iris)
lr <- lm_blog(Sepal.Length ~ Petal.Width + Petal.Length + Species,iris)
summary_blog(lr)
}
