% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcascore.R
\name{scoreplot}
\alias{scoreplot}
\title{PCA Score Plot}
\usage{
scoreplot(object, pc = c(1, 2), scale = 1, grouping = NULL)
}
\arguments{
\item{object}{a object of class \code{pca} generated using the \code{pca()} function.}

\item{pc}{a vector indicating the two principal components that should be plotted.}

\item{scale}{a number between 0 and 1 to scale the rotated data stored in \code{object}. The scaling procedure is the same as used in \code{autoplot()} for \code{prcomp} objects in \code{ggplot2}.}

\item{grouping}{a vector of the same length and in the same order as the data used in \code{pca()} that corresponds to categories of the original data.}
}
\value{
Returns a ggplot. This plot can be modified by adding additional ggplot functions.
}
\description{
Plots the rotated data from a \code{pca} object.
It represents the projection of the original data into the space spanned by the two specified principal components.
}
\examples{
require(ggplot2)
require(ggfortify)
require(patchwork)
# Compare to score plot from ggplot2/ggfortify
data(iris)
p1 = pca(iris[,-5]) # pca from pca package
p2 = prcomp(iris[,-5], scale=TRUE) # pca from stats package
plot1 = scoreplot(p1, pc=c(1,2), grouping=iris[,5])
plot2 = autoplot(p2, data=iris, colour='Species') # from ggfortify
plot1 + plot2

# Plot all combinations of PCs
data(swiss)
q = pca(~Fertility + Agriculture + Catholic + Infant.Mortality, data=swiss)
pcs = combn(x=1:4, m=2)
plot_list = vector("list", ncol(pcs))
for(i in 1:ncol(pcs)) {
    plot_list[[i]] = scoreplot(q, pc=pcs[,i])
}
wrap_plots(plot_list)

# Set scale to 0 to remove scaling
data(diamonds)
r = pca(~carat+depth+table+x+y+z, data=diamonds)
plot_scaled = scoreplot(r, scale=1)
plot_unscaled = scoreplot(r, scale=0)
plot_scaled + plot_unscaled # compare scaled and unscaled plots
}
