% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_i.R
\name{sig_i}
\alias{sig_i}
\title{Check significance of p-values}
\usage{
sig_i(pval)
}
\arguments{
\item{pval}{Array of p-values.}
}
\value{
\emph{sig_i} returns an array, where each element is the p value appended with a symbol denoting it's significant. The meaning of the symbols is as follows:\cr
\tabular{rrrrr}{
  \strong{Symbol} \tab \strong{Meaning}\cr
  \emph{***} \tab p-value<0.001\cr
  \emph{**} \tab p-value<0.01\cr
  \emph{*} \tab p-value<0.05\cr
  \emph{.} \tab p-value<0.1
}
}
\description{
\emph{sig_i} assigns a significance value to of p-value in an array of p-values.
}
\details{
sig_i takes an array containing p-values as an argument and assigns significance to each p value.
\emph{sig_i} is used to convert non numeric covariates into binary dummy variables
}
\examples{
p_array <- c(0.0004,0.06,0.5,1)
sig_i(p_array)
}
