% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-fit_model.R
\name{slr}
\alias{slr}
\title{slr}
\usage{
slr(outcome.str, covariate.str, data)
}
\arguments{
\item{outcome.str}{The string of the name of a numeric outcome variable in the dataset, e.g., 'Depression'}

\item{covariate.str}{The string of the name of a numeric input variable in the dataset, e.g., 'Age'}

\item{data}{A data frame containing the selected variables in the model.}
}
\value{
A seven-object list of the fitted model parameters, including the formula, a coefficient table, R-squared, F-statistics, degrees of freedom and the p-value.
}
\description{
Simple Linear Regression (SLR) is to estimate the effect of a single input variable on an interested outcome variable in an observed data set.
}
\details{
Can use the existing data \code{stroke} and select input and outcome variables to try this function out. Also remember that no NA should exist in the selected variables, otherwise the data needs imputation.
}
\examples{
\dontrun{
if(interactive()){
 fitted <- slr('Depression','Age',stroke)
 }
}
}
