% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_blog.R
\name{summary_blog}
\alias{summary_blog}
\title{Create Summary Table}
\usage{
summary_blog(lr)
}
\value{
Displays the following result from the model fit:\cr
Function call for \emph{lm_blog}\cr
Summary statistics of residuals\cr
Coefficient estimates and p-values\cr
Residual standard error\cr
Multiple R-squared & Adjusted R-squared\cr
}
\description{
\emph{summary_blog} displays the result summaries of the results of a linear regression model fitted using \emph{lm_blog}
}
\details{
\emph{summary_blog} generates a results summary using the \emph{lm_blog} returned list as an argument
}
\examples{
summary_blog(lr)
Where \emph{lr} is a fitted linear regeression model
}
